+++ pkg/analyzer/CHANGELOG.md
@@ -6,6 +6,8 @@
 * Deprecated `LibraryElement.isBrowserApplication`.
 * Deprecated `LibraryElement.libraryExports`,
   use `CompilationUnitElement.libraryExports` instead.
+* Deprecated `LibraryElement.libraryImports`,
+  use `CompilationUnitElement.libraryImports` instead.
 
 ## 6.9.0
 * `NormalFormalParameter` now implements `AnnotatedNode`.